+++ Mistward/Package/CHANGELOG.md
@@ -1,3 +1,30 @@
+ **0.7.0**
+---
+```
+- Adds support for live-configuration editing of crafting category
+- Updates configuration format to better support dynamic categories
+```
+
+ **0.6.1**
+---
+```
+- Better support for non-refundable costs.
+```
+
+ **0.6.0**
+---
+```
+- Enabled hot-reloading of structure configuration crafting costs and requirements.
+```
+
+ **0.5.1**
+---
+```
+- Readme updates
+- Reduce Localization added logging
+- Enabled hot-reloading of the demister range
+```
+
  **0.5.0**
 ---
 ```